oneMemoとCategoryIsMemoをmoleculeに移動
